testRun. error if delete not existing diff

+++ src/test-runs/test-runs.service.ts
@@ -116,7 +116,7 @@ export class TestRunsService {
 
     try {
       Promise.all([
-        this.staticService.deleteImage(testRun.diffName),
+        testRun.diffName && this.staticService.deleteImage(testRun.diffName),
         this.staticService.deleteImage(testRun.imageName),
       ]);
     } catch (err) {
